Fixed the LoggerTests fmt formatting issue
For conan, we install fmt 10.2.1 prebuilt conan package to not conflict with spdlog 1.14.0 prebuilt conan package.
However, when we switched to a newer version of fmt, that caused the LoggerTest to fail, which didn't happen
when we were using fmt 10.1.0 that was version installed when we used the standalone CMake approach. So, we updated
our TEST_CASE for 'fmt formatting works with the logger' to explicity create the chrono minutes duration,
followed by fmt format and then we pass our formatted minutes duration to our logger->log_critical, so we get
the expected fmt formatting to be '1m'. We didn't have to do this explicitly when we used fmt 10.1.0 since even
if we passed our logger->log_critical '1min', it would auto format to '1m'. For fmt 10.2.1 to have backward compatibility
with fmt 10.1.0 and the result we expected from this TEST_CASE in LoggerTests, we made this update.

I checked that when I build MiNiFi using conan and then build MiNiFi using standalone CMake, their LoggerTest PASSES.